requirements.txt
Flask==1.1.1   
jinja2<3.1.0
itsdangerous==2.0.1
werkzeug==2.0.2
boto3==1.22.4
mysql-connector-python==8.0.28
python-dotenv==0.19.2
